pyproject.toml
[tool.poetry]
name = "onener"
version = "0.1.0"
description = "Named Entity Recognition"
authors = ["Sansan, Inc."]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.8"
mecab-python3 = "^1.0"
pytorch-crf = "^0.7"
seqeval = "0.0.12"
torch = "~1.7.0"
transformers = "~3.2.0"
tqdm = "^4.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"